+++ src/data/bucket/line_bucket.test.ts
@@ -27,11 +27,11 @@ function createLineBucket({id, layout, paint, globalState, availableImages}: Cre
         type: 'line',
         layout,
         paint
-    } as LayerSpecification);
-    layer.recalculate({zoom: 0, zoomHistory: {} as ZoomHistory, globalState} as EvaluationParameters,
+    } as LayerSpecification, globalState);
+    layer.recalculate({zoom: 0, zoomHistory: {} as ZoomHistory} as EvaluationParameters,
         availableImages as Array<string>);
 
-    return new LineBucket({layers: [layer], globalState} as BucketParameters<LineStyleLayer>);
+    return new LineBucket({layers: [layer]} as BucketParameters<LineStyleLayer>);
 }
 
 describe('LineBucket', () => {