+++ src/source/worker_tile.ts
@@ -37,7 +37,6 @@ export class WorkerTile {
     showCollisionBoxes: boolean;
     collectResourceTiming: boolean;
     returnDependencies: boolean;
-    globalState: Record<string, any>;
 
     status: 'parsing' | 'done';
     data: VectorTile;
@@ -60,7 +59,6 @@ export class WorkerTile {
         this.returnDependencies = !!params.returnDependencies;
         this.promoteId = params.promoteId;
         this.inFlightDependencies = [];
-        this.globalState = params.globalState;
     }
 
     async parse(data: VectorTile, layerIndex: StyleLayerIndex, availableImages: Array<string>, actor: IActor, subdivisionGranularity: SubdivisionGranularitySetting): Promise<WorkerTileResult> {
@@ -115,7 +113,7 @@ export class WorkerTile {
                 if (layer.maxzoom && this.zoom >= layer.maxzoom) continue;
                 if (layer.visibility === 'none') continue;
 
-                recalculateLayers(family, this.zoom, availableImages, this.globalState);
+                recalculateLayers(family, this.zoom, availableImages);
 
                 const bucket = buckets[layer.id] = layer.createBucket({
                     index: featureIndex.bucketLayerIDs.length,
@@ -125,8 +123,7 @@ export class WorkerTile {
                     overscaling: this.overscaling,
                     collisionBoxArray: this.collisionBoxArray,
                     sourceLayerIndex,
-                    sourceID: this.source,
-                    globalState: this.globalState
+                    sourceID: this.source
                 });
 
                 bucket.populate(features, options, this.tileID.canonical);
@@ -181,7 +178,7 @@ export class WorkerTile {
         for (const key in buckets) {
             const bucket = buckets[key];
             if (bucket instanceof SymbolBucket) {
-                recalculateLayers(bucket.layers, this.zoom, availableImages, this.globalState);
+                recalculateLayers(bucket.layers, this.zoom, availableImages);
                 performSymbolLayout({
                     bucket,
                     glyphMap,
@@ -196,7 +193,7 @@ export class WorkerTile {
                 (bucket instanceof LineBucket ||
                 bucket instanceof FillBucket ||
                 bucket instanceof FillExtrusionBucket)) {
-                recalculateLayers(bucket.layers, this.zoom, availableImages, this.globalState);
+                recalculateLayers(bucket.layers, this.zoom, availableImages);
                 bucket.addFeatures(options, this.tileID.canonical, imageAtlas.patternPositions, lineAtlas);
             }
         }
@@ -216,11 +213,10 @@ export class WorkerTile {
     }
 }
 
-function recalculateLayers(layers: ReadonlyArray<StyleLayer>, zoom: number, availableImages: Array<string>, globalState: Record<string, any>) {
+function recalculateLayers(layers: ReadonlyArray<StyleLayer>, zoom: number, availableImages: Array<string>) {
     // Layers are shared and may have been used by a WorkerTile with a different zoom.
     const parameters = new EvaluationParameters(zoom);
     for (const layer of layers) {
-        layer.setGlobalState(globalState);
         layer.recalculate(parameters, availableImages);
     }
 }